Fixed steal identity. Removed duplicated items in globals.hpp. Renamed util/player to util/local_player.
