+++ BigBaseV2/src/util/ped.hpp
@@ -25,16 +25,15 @@ namespace big::ped
 
 	inline void steal_identity(const Ped target)
 	{
-		Ped ped = self::ped;
-
-		const int max_health = ENTITY::GET_ENTITY_MAX_HEALTH(ped);
-		const int current_health = ENTITY::GET_ENTITY_HEALTH(ped);
-		const int current_armor = PED::GET_PED_ARMOUR(ped);
+		const int max_health = ENTITY::GET_ENTITY_MAX_HEALTH(self::ped);
+		const int current_health = ENTITY::GET_ENTITY_HEALTH(self::ped);
+		const int current_armor = PED::GET_PED_ARMOUR(self::ped);
 
 		PLAYER::SET_PLAYER_MODEL(self::id , ENTITY::GET_ENTITY_MODEL(target));
-		PED::CLONE_PED_TO_TARGET(target, ped);
-		ENTITY::SET_ENTITY_MAX_HEALTH(ped, max_health);
-		ENTITY::SET_ENTITY_HEALTH(ped, current_health, 0);
-		PED::SET_PED_ARMOUR(ped, current_armor);
+		script::get_current()->yield();
+		PED::CLONE_PED_TO_TARGET(target, self::ped);
+		ENTITY::SET_ENTITY_MAX_HEALTH(self::ped, max_health);
+		ENTITY::SET_ENTITY_HEALTH(self::ped, current_health, 0);
+		PED::SET_PED_ARMOUR(self::ped, current_armor);
 	}
 }
